Separate local and azure wrappers for fetch_universe_metadata job
