+++ jobs/fetch-universe-metadata/save_metadata_azure/function.json
